+++ config/routes.rb
@@ -1,5 +1,7 @@
 Rails.application.routes.draw do
-  resources :accounts, :except => [:new, :edit], :defaults => {:format => 'json'}
-  get '/settlement' => 'accounts#settle', :defaults => {:format => 'json'}
-  get '/' => 'accounts#manage'
+  resources :payments, :except => [:new, :edit], :defaults => {:format => 'json'}
+  get '/settlement' => 'payments#settle', :defaults => {:format => 'json'}
+  get '/' => 'payments#manage'
+  get '/login' => 'login#form'
+  post '/login' => 'login#authenticate_user'
 end