+++ db/migrate/20161105070905_create_clients.rb
@@ -1,0 +1,10 @@
+class CreateClients < ActiveRecord::Migration
+  def change
+    create_table :clients do |t|
+      t.string :application_id
+      t.string :application_key
+
+      t.timestamps
+    end
+  end
+end