+++ db/migrate/20161211042729_rename_accounts_to_payments.rb
@@ -1,0 +1,5 @@
+class RenameAccountsToPayments < ActiveRecord::Migration
+  def change
+    rename_table :accounts, :payments
+  end
+end